requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
